Implement Emote command with buttons
